Add project setup files and initial script
This commit introduces several files to set up the project environment and manage dependencies. The 'Makefile' contains a 'conda_export' target that will export the environment setup to 'requirements.txt'. The '.gitignore' file is added to exclude '.idea' directory from the repo. 'README.md' is updated with commands to setup, activate, and update Conda environment. An 'environment.yaml' file is created to specify the Conda environment including all required dependencies. The 'Documents.iml' is for IDE settings. And, finally, an initial 'main.py' script is added as a starting point for the project.